requirements.txt
# Core data processing
pandas>=1.5.0
polars>=0.19.0
numpy>=1.21.0

# Elasticsearch integration
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0

# Survival analysis
lifelines>=0.27.0

# Statistical analysis
scipy>=1.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration management
python-dotenv>=0.19.0

# Optional: Performance optimization
# fastparquet>=0.8.0  # For faster parquet I/O
# pyarrow>=10.0.0     # Alternative parquet engine